Fix race condition resulting incorrect motor target
- There was an issue with commands issued in quick succession, combined
with the ability of the motor turn ISR to reset the target position to 0
that caused steps to be lost in certain circumstances
- The ISR has been minimized and now does not manipulate state outside
of the motor object
- target_turns is now never reset except when it is about to overflow
(appox 330k turns are required for this to happen). This will result in
an an icorrect position but I think we can live with this.
- More diagnostic info added to {print:} output which shows the current
absolute target, the maximum number of turns allowed, etc